Close sheet on escape
When a sheet was using in an applet configuration the escape key would
close the dialog rather than the modal sheet.

There were two issues, we need to intercept escape being used as a
shortcut for other actions, such as closing the parent dialog.

contentItemParent is the item set to have focus so key handling needs to
be there.

BUG: 430563